Enhance SEO metadata and content
