+++ app/layout.tsx
@@ -2,10 +2,76 @@ import type { Metadata } from 'next'
 import Link from 'next/link'
 import './globals.css'
 
+const siteUrl = 'https://ippan.org'
+
 export const metadata: Metadata = {
-  title: 'IPPAN Foundation',
+  metadataBase: new URL(siteUrl),
+  title: {
+    default: 'IPPAN Foundation',
+    template: '%s | IPPAN Foundation',
+  },
   description:
-    'IPPAN Foundation stewards the IPPAN Network—an energy-efficient, massively scalable public blockchain built for real-world adoption.',
+    'IPPAN Foundation stewards the IPPAN Network—an energy-efficient, massively scalable public blockchain built for real-world adoption and responsible decentralization.',
+  keywords: [
+    'IPPAN Foundation',
+    'IPPAN Network',
+    'blockchain foundation',
+    'decentralized infrastructure',
+    'energy efficient blockchain',
+    'web3 non-profit',
+  ],
+  openGraph: {
+    type: 'website',
+    url: siteUrl,
+    siteName: 'IPPAN Foundation',
+    title: 'IPPAN Foundation',
+    description:
+      'The IPPAN Foundation advances an energy-efficient, massively scalable public blockchain for real-world adoption, governance, and ecosystem growth.',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    creator: '@ippan',
+    title: 'IPPAN Foundation',
+    description:
+      'Explore the IPPAN Foundation—stewards of the IPPAN Network and advocates for sustainable decentralized infrastructure.',
+  },
+  alternates: {
+    canonical: siteUrl,
+  },
+  robots: {
+    index: true,
+    follow: true,
+  },
+}
+
+const organizationStructuredData = {
+  '@context': 'https://schema.org',
+  '@type': 'Organization',
+  name: 'IPPAN Foundation',
+  url: siteUrl,
+  description:
+    'The IPPAN Foundation is a non-profit stewarding the IPPAN Network—an energy-efficient, scalable blockchain for real-world adoption.',
+  sameAs: [
+    'https://x.com/ippan',
+    'https://linkedin.com/company/ippan-foundation',
+    'https://github.com/ippan-foundation',
+  ],
+  contactPoint: [
+    {
+      '@type': 'ContactPoint',
+      email: 'info@ippan.org',
+      contactType: 'general',
+      areaServed: 'Global',
+      availableLanguage: ['English'],
+    },
+    {
+      '@type': 'ContactPoint',
+      email: 'press@ippan.org',
+      contactType: 'press',
+      areaServed: 'Global',
+      availableLanguage: ['English'],
+    },
+  ],
 }
 
 const navLinks = [
@@ -28,7 +94,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
               <Link href="/" className="text-xl font-semibold tracking-tight">
                 IPPAN Foundation
               </Link>
-              <nav className="flex items-center gap-6 text-sm font-medium text-gray-600">
+              <nav aria-label="Primary" className="flex items-center gap-6 text-sm font-medium text-gray-600">
                 {navLinks.map((link) => (
                   <Link key={link.href} href={link.href} className="transition hover:text-gray-900">
                     {link.label}
@@ -47,6 +113,11 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
             </div>
           </footer>
         </div>
+        <script
+          type="application/ld+json"
+          suppressHydrationWarning
+          dangerouslySetInnerHTML={{ __html: JSON.stringify(organizationStructuredData) }}
+        />
       </body>
     </html>
   )
